feat: Add isSuccess property to UploadFineTuningExamplesResponse
This commit adds the `isSuccess` property to the `UploadFineTuningExamplesResponse` class in order to indicate whether the upload of fine-tuning examples was successful or not. The property is set to `true` in the `uploadFineTuningExamples()` method of the `RestFineTuningDataController`. Additionally, some test cases and feature scenarios were updated to reflect this change.