feat: Add scenario for editing note details with Shift-Enter line break support
This commit introduces a new scenario in the note edit feature to test the functionality of adding line breaks in note details using Shift-Enter. It also updates the notePage object to handle the parsing of special keys and verifies the presence of line breaks in the note details during tests.